MDL-51458 testing: Disable Travis notifications entirely
The Travis-CI service currently over-notifies the upstream repository
owner which can mean that committers without access to a fork of the
repository are notified of failures, and success changes.

Until this issue is addressed, and we are happy with the stability of our
build, we have disabled notification support.

The Travis-CI service problem has been raised in
https://github.com/travis-ci/travis-ci/issues/4976