allow import of password-protected calendars
Calendar subscriptions of password-protected calendars are not possible as of 2015-11-23. This is due to URLs of the form user:password@example.com/myCalendar.ics being rejected as invalid. If these were accepted, which the proposed change makes sure, subscribing to pw-protected calendars would be possible. As curl can already handle such URLs, no further changes are required.

Downside: I do not know what other parts of Moodle would be affected in what ways by this change (to a very central function).