MDL-52718 block_community: Send XML-RPC request correctly
mnet has its own separate XML-RPC implementation which supports
introspection. So even though we were sending incorrectly formatted
parameters, our previous XML-RPC implementation could work out how
to mangle the parameters to be correct. Introspection is not a
part of the XML-RPC specification and we should not rely on it. Our
current XML-RPC client/server implementation does not support it either.

Therefore we should always send data in the expected format.