MDL-53595 install: update the list of stale files for 3.1
Note I've deleted a couple of files there, 99.99% not problematic
(repository/boxnet) because as a general rule it's better to avoid
adding plugin files in the function and, instead, use core ones that
cannot be "reinstalled". Exception to the rule (what's a rule without
an exception) being forum that cannot be deleted. ;-)