MDL-54121 core-notification: Make dialogues re-calculate the lockscroll
When the content of a dialogue changes after it is first opened - you need to call centerDialogue()
to make sure it still sits in the middle of the window. In addition this now recalculates the "lockScroll"
so that if you need to scroll down to see all of the dialogue you can do it.