MDL-54649 quiz/question: fix css lint errors
Mostly the rules were not having any effect so have been removed.

The text-wrap rule was removed rather than corrected because overflow
was clipped with ellipsis anyway.

An exception for !important has been added to mod/quiz/styles.css
because the use in the secure style is by design. But the exception
isn't ideal because it applied to the whole file and csslint doesn't
seem to be very intelligent at appling line rules at a more granular
level at this point in time.