MDL-56178 theme_boost: Display drop zones labels
The markup has changed so we now need to target .sr-only rather
than just .accesshide. I did keep the latter in case it's still used
somewhere, in a custom form field for instance.

The ideal solution was to tell QuickForm to display the labels of
those elements in the group. That works for Boost, but that did not
work for Clean. The rendering mechanics of QuickForm decides to not
print the label, unless it's set as hidden... the CSS hack remains.