MDL-53695 lib/javascript.php: ensure full env for core_minify
This applies the same approach to generating minified js as is already
used in theme/javascript.php and other scripts. When we need to do the
minification, we load the full Moodle environment, so we can use the
libraries.

The reason this was required was because now the minifier uses
autoloaded classes, which are not autoloaded when doing
ABORT_AFTER_CONFIG.

Fred persuaded me that it's probably safer to keep it that way, rather
than alllow ABORT_AFTER_CONFIG scripts from autoloading the world and
their dog (especially for resource serving scripts, they are intended to
be lightweight when serving from cache).