MDL-56509 tool_usertours: Remove guest from list and add admin
It is not possible to display a tour to users with the guest role so remove
all users with this archetype from the list.

This also changes behaviour such that it is possible to target a tour just
to administrators and administrators are not shown the tours for roles
which they do not have in the current context.