MDL-56601 core_user: Disable profile picture field if user is not setup
Sometimes it is possible that a user is presented with the profile
edit form, but is in a state where they are not considered fully
set up (one example is when a site is installed using the CLI script)

In these cases the filepicker gets upset when the user tries
to upload a profile picture.

This patch simply disables that field in these cases and replaces it
with an error message explaining what's wrong.