MDL-56768 boost: Special positioning for grader report
Grader report has custom overflow handling which does not cope with a
scrollable content region. So we need to make the content region flow
weirdly out to the right of the page, and push the blocks underneath -
even on really wide screens where it looks bad. We also needed to
upgrade the custom scroll code to handle the margin on the body
changing.