MDL-41729 auth_shibboleth: Allow password change for shibboleth users
This patch modifies can_change_password() and adds change_password_url() to allow Shibboleth users to change their password
in case there is an (external) password change URL defined in Moodle. If no such URL is defined, the behaviour is the same
as without the proposed change.

Thanks FH-HWZ.ch for the contribution.