MDL-56800 theme_boost: Display frozen form's radios as disabled
For frozen (read-only) forms, the radio widget was not displayed at all
and only the legend text was displayed. So it was not possible to see
which one of the radios in a group was selected.

This patch mimics the original quick form rendering and displays the
radio as disabled.